fixes errors due to too long command line parameters
- This fix helps adressing an issue many people are having in #36:
- When C++ sourcecode like e.g. the one from notepad++ is used with
include-what-you-use the commands are too long for clang which is behind iwyu.
For these reasons clang supports so called "support files". These files
contain the complete clang file list. The clang parameters and IWYU
parameters need still to be given on the command line itself.
- A support file is given as @filename. It contains all includes, defines,
and the filename to work on.
- Apart from using support files this fix also increases readability by:
-- removes unnecessary scope
-- better orders variable definitions to where they are used
-- create lists of clang options and iwyu options. These lists are later
on assembled to the correct parameter string and formerly redundant parameter
prefixes are happening only once (-Xiwyu --parameter1, -Xiwyu
--parameter2, ...)